models/video_editors/text2vidzero/requirements.txt
accelerate==0.16.0
addict==2.4.0
albumentations==1.3.0
basicsr==1.4.2
decord==0.6.0
diffusers==0.14.0
einops==0.6.0
gradio==3.23.0
kornia==0.6
imageio==2.9.0
imageio-ffmpeg==0.4.2
invisible-watermark>=0.1.5
moviepy==1.0.3
numpy==1.24.1
omegaconf==2.3.0
open_clip_torch==2.16.0
opencv_python==4.7.0.72
opencv-contrib-python==4.7.0.72
Pillow==9.4.0
pytorch_lightning==1.5.0
prettytable==3.6.0
scikit_image==0.19.3
scipy==1.10.1
tensorboardX==2.6
torchmetrics==0.6.0
tqdm==4.64.1
timm==0.6.12
transformers==4.26.0
test-tube>=0.7.5
webdataset==0.2.5
yapf==0.32.0
safetensors==0.2.7
beautifulsoup4
bs4
tomesd
